# openapi.yml
openapi: 3.0.3
info:
  title: Fineract CN Amun
  version: 1.0.0
  description: An API module for credit risk assessment
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/v1/algorithms/predict:
    post:
      operationId: algorithms_predict
      description: Predict credit risk for a loan
      parameters:
      - in: query
        name: status
        schema:
          type: string
        description: The status of the algorithm
        examples:
          Example1:
            value: production
            summary: Algorithm in production
      - in: query
        name: version
        schema:
          type: string
        description: Algorithm version
        required: true
      tags:
      - algorithms
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Input Model Data'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Input Model Data'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Input Model Data'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoring successful'
          description: ''
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: Session
servers:
- url: http://127.0.0.1:8000
  description: server on localhost